Extract StagingAreaValidator class to a python package for use as a library, factor out argparse code to the top-level script Add README, setup.cfg and pyprojec.toml for library packaging
