+++ README.md
@@ -1,0 +1,4 @@
+# HCA Staging Import Validator
+
+Runs a pre-check of a staging area to identify issues that might cause the
+snapshot or indexing processes to fail.
