+++ setup.cfg
@@ -1,0 +1,13 @@
+[metadata]
+name = hca-import-validation
+version = 0.0.1
+description = HCA Staging Import Validation
+url = https://github.com/dataBiosphere/hca-import-validation
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: Apache Software License
+    Operating System :: OS Independent
+
+[options]
+packages = find:
+python_requires = >=3.6
